chore(mappings): write mappings for registry
